Add files via upload
FORECASTLY_V6 is just a name, this is not to predict the future values of stocks rather only to manage the stocks (automated). Apologies for the misleading name

The excel files provided are only for reference, with some data to run the code on, and it serves as a format for your excel files. All you need to do if you are making your own, is enter the link of the stocks you want to manage from screener.in as it has realiable data. The reason this works with link instead of NSE tag or stock name is some stocks have a variant which tracks the parent company instead of the individual company which leads to certain problems and confusions.